requirements.txt
asgiref==3.7.2
Django==5.0.1
django-appconf==1.0.6
django-imagekit==5.0.0
mysqlclient==2.2.4
pilkit==3.0
pillow==10.2.0
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4